log formatting error

+++ src/main/java/gr/knowledge/internship/introduction/repository/EmployeeRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import gr.knowledge.internship.introduction.entity.Employee;
 
 public interface EmployeeRepository extends JpaRepository<Employee, Long> {
-	public List<Employee> findByCompanyId(int companyId);
+	public List<Employee> findByCompanyId(Long companyId);
 
 	public List<Employee> getEmployeeByCompanyId(Long companyId);
 }
